+++ CHANGELOG.md
@@ -7,7 +7,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
-- [Migrate modals to side panel](https://github.com/multiversx/mx-sdk-dapp-core/pull/122)
+- [Fixed toasts completion and notifications feed toasts management](https://github.com/multiversx/mx-sdk-dapp-core/pull/131)
 - [Toasts data refactoring](https://github.com/multiversx/mx-sdk-dapp-core/pull/129)
 - [Added ledger idle connection manager](https://github.com/multiversx/mx-sdk-dapp-core/pull/127)
 - [Fixed ledger logout on page reload](https://github.com/multiversx/mx-sdk-dapp-core/pull/126)